Applied adviced changes
